moved updateCloudMetadata and processFile fuctions from index.js to empolis_ops.js
