+++ index.js
@@ -1,11 +1,9 @@
 // Imports
 import { loadConfig, setConfig, getConfig } from './config.js';
-import { getToken } from './empolis_admin.js';
 import { fileSearch } from './empolis_search.js';
-import { editFileMetadata } from './empolis_ops.js';
-import { readJsonData } from './helpers.js';
+import { updateCloudMetadata } from './empolis_ops.js';
 import { createUpdateIndexFile } from './index_creation.js';
-import logger, { configureLogger, logPrettyJson } from './logger.js';
+import logger, { configureLogger } from './logger.js';
 import { performance } from 'node:perf_hooks';
 import { fileURLToPath } from 'url';
 import path from 'node:path';
@@ -94,105 +92,6 @@ async function main() {
       `Total execution time: ${chalk.cyan(executionTime.toFixed(2))} seconds.` +
         ` Logs can be found in ${chalk.cyan(logger.transports[0].dirname)}.`
     );
-  }
-}
-
-/**
- * Update the metadata of files in the Empolis cloud with information from the index file
- * @async
- * @function updateCloudMetadata
- * @memberof empolis_ops
- * @param {Array} fileList - list of filenames to update metadata for
- * @param {string} indexFile - path of the index file containing metadata
- * @requires empolis_admin
- * @returns {Promise<null>} null
- */
-async function updateCloudMetadata({ fileList, indexFile }) {
-  const config = getConfig();
-  const API_TOKEN = await getToken();
-  // Load the full index of files from the index file
-  const index = await readJsonData(indexFile);
-  // Update the metadata for each file in the index
-  console.log(`  Updating the metadata of ${chalk.cyan(fileList.length)} files...`);
-  logger.info(`Updating the metadata of ${fileList.length} files...`);
-  for (const file of fileList) {
-    const fileIndex = index.findIndex((obj) => obj.filename === file);
-    const fileData = index[fileIndex];
-    await processFile(API_TOKEN, fileData);
-  }
-  console.log(
-    `${chalk.green('√')}` +
-      ` Completed metadata update operation for '${config.dataSourceSelection}' data source`
-  );
-  logger.info(
-    `Completed metadata update operation for '${config.dataSourceSelection}' data source`
-  );
-  return null;
-}
-
-/**
- * Function to modify the metadata of a file via the Empolis INGEST API
- * <br> Only modifies metadata with editFileMetadata() if it does not have the correct value already
- * @async
- * @function processFile
- * @param {string} authToken - authentication token for API requests
- * @param {string} file - filename of the file to process
- * @param {string} htmlTitle - title extracted from the HTML file
- * @requires ./empolis_functions.js
- * @requires ./helpers.js
- * @returns nothing
- */
-
-async function processFile(authToken, dataObject) {
-  logger.info(`Processing ${dataObject.filename}`);
-  logger.debug(`dataObject: ${JSON.stringify(dataObject)}`);
-
-  // Search for file in the user selected data source to get the current metadata
-  const fileMetadata = await fileSearch({ searchTerm: dataObject.filename, consoleOutput: false });
-
-  let newKeywords = '';
-  if (dataObject.breadcrumbs) {
-    for (const keyword of dataObject.breadcrumbs) {
-      newKeywords = newKeywords + keyword + '; ';
-    }
-    newKeywords = newKeywords.trim();
-  }
-
-  // Check if 'Title' and 'Keywords' are already correct
-  if (fileMetadata.Title && dataObject.title) {
-    if (fileMetadata.Title.toLowerCase() === dataObject.title.toLowerCase()) {
-      logger.info(`${dataObject.filename} already has the correct title`);
-      // Return if keywords and title already match
-      if (fileMetadata.Keywords_txt) {
-        if (fileMetadata.Keywords_txt.toLowerCase() === newKeywords.toLowerCase()) {
-          logger.info(
-            `${dataObject.filename} already has the correct title and keywords, metadata will not be updated`
-          );
-          return;
-        }
-      }
-      // Return if title matches, entry has no keywords, and there are no new keywords to be added
-      else if (newKeywords.length === 0) {
-        logger.info(
-          `Keywords for ${dataObject.filename} do not exist and title is correct, metadata will not be updated`
-        );
-        return;
-      }
-    }
-  }
-
-  // Build new metadata object with Title and optional Keywords
-  let newMetadata = { ...fileMetadata, Title: dataObject.title };
-  if (newKeywords.length > 0) {
-    newMetadata = { ...newMetadata, Keywords_txt: newKeywords };
-  }
-  logPrettyJson(newMetadata, 'newMetadata');
-
-  const editMetadataResponse = await editFileMetadata({ newMetadata });
-  if (editMetadataResponse === 202) {
-    logger.info(`${dataObject.filename} metadata modified successfully`);
-  } else {
-    logger.error(`Failed to modify metadata for ${dataObject.filename}`);
   }
 }
